--- merge_mvc/GugudanView.java
@@ -1,12 +1,0 @@
-package merge_mvc;
-
-	public class GugudanView {
-	
-		public static void main(String[] args) {
-			
-			System.out.println("1 ~ 999 사이의 단수를 입력하세요 : "); // 입력하라는 문구
-			
-			GugudanModel.GugudanInput();
-			
-		}
-}
