+++ merge_mvc/MergePracticeView.java
@@ -10,9 +10,10 @@ class MergePracticeView {
 	// 메뉴 표시
 	public void showMenu() {
 		System.out.println("<실행할 작업>");
-		System.out.println("1 : 구구단");
-		System.out.println("2 : 계산기");
-		System.out.println("3 : 별찍기");
+		System.out.println("1 : 구구단 ver.1");
+		System.out.println("2 : 구구단 ver.2");
+		System.out.println("3 : 계산기 ver.1");
+		System.out.println("4 : 계산기 ver.2");
 		System.out.println("0 : 종료");
 		System.out.println("----------");
 	}
