+++ merge_mvc/gugudan1/GugudanView.java
@@ -1,0 +1,12 @@
+package merge_mvc.gugudan1;
+
+	public class GugudanView {
+	
+		public static void main(String[] args) {
+			
+			System.out.println("1 ~ 999 사이의 단수를 입력하세요 : "); // 입력하라는 문구
+			
+			GugudanModel.gugudanInput();
+			
+		}
+}
